Revert " Implemented some operations in the webservice authorization."
