+++ client/src/components/root/navbar.vue
@@ -17,7 +17,6 @@
     methods: {
       ...mapActions(['setToken', 'setUser']),
       logout() {
-        this.$http.post('auth/revoke');
         this.setToken('')
         this.setUser({})
         this.$router.push({ name: 'login.index' })
