+++ webservice/app/Http/Controllers/ApiController.php
@@ -81,19 +81,6 @@ abstract class ApiController extends Controller
     }
 
     /**
-     * Return a 204 response.
-     *
-     * @param  string $message
-     *
-     * @return \Illuminate\Http\Response
-     */
-    protected function responseWithNoContent()
-    {
-        return $this->setStatusCode(Response::HTTP_NO_CONTENT)
-                    ->response([]);
-    }
-
-    /**
      * Return an error response.
      *
      * @param  mixed $message
