+++ webservice/app/Http/Controllers/LoginController.php
@@ -4,10 +4,11 @@ namespace App\Http\Controllers;
 
 use Illuminate\Foundation\Auth\ThrottlesLogins;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Auth;
 use Lang;
 
-class AuthController extends ApiController
+class LoginController extends ApiController
 {
     use ThrottlesLogins;
 
@@ -18,7 +19,7 @@ class AuthController extends ApiController
      * @param Request $request
      * @return \Illuminate\Http\JsonResponse
      */
-    public function issueToken(Request $request)
+    public function login(Request $request)
     {
         // Determine if the user has too many failed login attempts.
         if ($this->hasTooManyLoginAttempts($request)) {
@@ -65,7 +66,7 @@ class AuthController extends ApiController
     /**
      * Return error message after determining invalid credentials.
      *
-     * @param  \Illuminate\Http\Request $request
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\JsonResponse
      */
     protected function sendFailedLoginResponse(Request $request)
@@ -78,7 +79,7 @@ class AuthController extends ApiController
     /**
      * Redirect the user after determining they are locked out.
      *
-     * @param  \Illuminate\Http\Request $request
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\JsonResponse
      */
     public function sendLockoutResponse(Request $request)
@@ -90,31 +91,6 @@ class AuthController extends ApiController
         $message = Lang::get('auth.throttle', ['seconds' => $seconds]);
 
         return $this->responseWithTooManyRequests($message);
-    }
-
-    /**
-     * Revoke the user's token.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function revokeToken()
-    {
-        Auth::guard('api')->logout();
-
-        return $this->responseWithNoContent();
-    }
-
-    /**
-     * Refresh the user's token.
-     *
-     * @param Request $request
-     * @return \Illuminate\Http\JsonResponse
-     */
-    public function refreshToken(Request $request)
-    {
-        $token = Auth::guard('api')->refresh();
-
-        return $this->response(compact('token'));
     }
 
     public function username()
