+++ webservice/routes/api.php
@@ -3,13 +3,11 @@
 Route::group([
     'middleware' => ['cors', 'api'],
 ], function () {
-    Route::post('/auth/issue', 'AuthController@issueToken');
-    Route::post('/auth/refresh', 'AuthController@refreshToken');
+    Route::post('/login', 'LoginController@login');
 
     Route::group([
-        'middleware' => 'jwt.auth',
+        'middleware' => 'auth:api',
     ], function () {
-        Route::post('/auth/revoke', 'AuthController@revokeToken');
         Route::resource('/categories', 'CategoriesController', [
             'except' => ['create', 'edit'],
         ]);
